requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.3.0
backoff==2.2.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
fastembed==0.4.2
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langfuse==2.54.1
langsmith==0.1.147
loguru==0.7.2
MarkupSafe==2.1.5
marshmallow==3.22.0
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
onnx==1.17.0
onnxruntime==1.19.2
openai==1.59.3
orjson==3.10.13
packaging==24.2
pandas==2.0.3
pillow==10.4.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
py_rust_stemmers==0.1.3
pydantic==2.9.2
pydantic_core==2.23.4
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.0
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
win32-setctime==1.1.0
wrapt==1.17.0
yarl==1.15.2
zipp==3.20.2
